feat(Forms): add `transformLabel` to Value.Composition
