+++ packages/dnb-eufemia/src/extensions/forms/Value/Composition/CompositionDocs.ts
@@ -1,10 +1,11 @@
 import { PropertiesTableProps } from '../../../../shared/types'
 import { ValueProperties } from '../ValueDocs'
 
-const { label } = ValueProperties
+const { label, transformLabel } = ValueProperties
 
 export const CompositionProperties: PropertiesTableProps = {
   label,
+  transformLabel,
   maxWidth: {
     doc: 'Use `small`, `medium` or `large` for predefined standard max widths. Defaults to `auto`.',
     type: 'string',
