+++ packages/dnb-eufemia/src/extensions/forms/Value/SummaryList/SummaryListDocs.ts
@@ -1,14 +1,12 @@
 import { PropertiesTableProps } from '../../../../shared/types'
+import { ValueProperties } from '../ValueDocs'
+
+const { transformLabel } = ValueProperties
 
 export const SummaryListProperties: PropertiesTableProps = {
   layout: {
     doc: 'Use `grid`, `horizontal` or `vertical`.',
     type: 'string',
-    status: 'optional',
-  },
-  transformLabel: {
-    doc: 'Transforms the label before it gets displayed. Receives the label as the first parameter. The second parameter is a object containing the `convertJsxToString` function.',
-    type: 'function',
     status: 'optional',
   },
   inheritVisibility: {
@@ -21,6 +19,7 @@ export const SummaryListProperties: PropertiesTableProps = {
     type: 'boolean',
     status: 'optional',
   },
+  transformLabel,
   children: {
     doc: 'Contents.',
     type: 'React.Node',
