+++ packages/dnb-eufemia/src/extensions/forms/hooks/useValueProps.ts
@@ -6,17 +6,12 @@ import {
   useRef,
 } from 'react'
 import { Path, ValueProps } from '../types'
-import { convertJsxToString } from '../../../shared/component-helper'
 import useExternalValue from './useExternalValue'
 import usePath from './usePath'
 import DataContext from '../DataContext/Context'
 import ValueProviderContext from '../Value/Provider/ValueProviderContext'
 
 export type Props<Value> = ValueProps<Value>
-
-const transformLabelParameters = {
-  convertJsxToString,
-} as unknown as Parameters<Props<unknown>['transformLabel']>[1]
 
 export default function useValueProps<
   Value = unknown,
@@ -34,7 +29,6 @@ export default function useValueProps<
     defaultValue,
     inheritVisibility,
     inheritLabel,
-    transformLabel = (label: Props['label']) => label,
     transformIn = (value: Value) => value,
     toInput = (value: Value) => value,
     fromExternal = (value: Value) => value,
@@ -93,11 +87,9 @@ export default function useValueProps<
     ? transformIn(toInput(externalValue))
     : undefined
 
-  const label = transformLabel(
+  const label =
     props.label ??
-      (inheritLabel ? fieldPropsRef?.current?.[path]?.label : undefined),
-    transformLabelParameters
-  )
+    (inheritLabel ? fieldPropsRef?.current?.[path]?.label : undefined)
 
   return { ...props, label, value }
 }
